Rebase fixes

+++ GameAPI/src/Monster.cpp
@@ -18,7 +18,7 @@ void Monster::AddTarget(int targetPos)
 
 void Monster::Activate(HeroesTeam& heroTeam)
 {
-	for (int i = 0; i < m_targets.size(); i++)
+	for (size_t i = 0; i < m_targets.size(); i++)
 	{
 		//m_monsterCard.GetAction().ApplyEffect(heroTeam.GetHero(m_targets.at(i)));
 	}
